Move the examples into an appendix
I softened the text with the discussion about handling segments. This is one of the comments from J. Touch. 
I moved most of the figures into an appendix to indicate these are implementation examples. 
I also added a reference to RFC7857 where we have a state machine for handling TCP connections. This is IMHO sufficient.